+++ src/renderer/utils/chart-layout/index.ts
@@ -20,11 +20,10 @@ export type {
 
 export { BOX_W, MIN_BOX_H, V_GAP, H_GAP, GEN_GAP, PAD, PORTRAIT_W, PORTRAIT_H, BOX_PAD_Y, BOX_PAD_X_LEFT, PORTRAIT_GAP, BOX_PAD_X_RIGHT, CURVE_R, TEXT_AREA_W, ADD_BTN_AREA_W } from './constants';
 export { wrapName, wrapFullNameSegments, measureBoxHeight } from './measure';
-export { curvedElbow, dashForSubtype } from './connectors';
-export type { ParentSubtypeForDash } from './connectors';
+export { curvedElbow } from './connectors';
 export { yearFromDate, maxDescendantDepth } from './utils';
 export { computePedigreeLayout } from './pedigree';
-export { computeHourglassLayout, maxDescendantDepthTP, FOSTER_PATH_PREFIX } from './hourglass';
+export { computeHourglassLayout, maxDescendantDepthTP } from './hourglass';
 export { computeTimelineLayout, eventSymbol } from './timeline';
 export { computeDescendantLayout } from './descendant';
 export { buildHourglassTree, buildPedigreeTreePerson, buildDescendantTreePerson, injectOutlines, findPerson, PLACEHOLDER_PREFIX } from './hourglass-tree';